+++ wisdom/platform/windows/wisdom/wisdom_windows.cpp
@@ -6,17 +6,17 @@
 #include <d3d11.h>
 
 namespace wis::detail {
-inline void ToSwapchainDesc(DXGI_SWAP_CHAIN_DESC1& swap_desc, const wis::SwapchainDesc* desc) noexcept
+inline void ToSwapchainDesc(DXGI_SWAP_CHAIN_DESC1& swap_desc, const wis::SwapchainDesc& desc) noexcept
 {
-    swap_desc.Width = desc->size.width;
-    swap_desc.Height = desc->size.height;
-    swap_desc.Format = convert_dx(desc->format);
-    swap_desc.Stereo = desc->stereo;
+    swap_desc.Width = desc.size.width;
+    swap_desc.Height = desc.size.height;
+    swap_desc.Format = convert_dx(desc.format);
+    swap_desc.Stereo = desc.stereo;
     swap_desc.SampleDesc.Count = 1u;
     swap_desc.SampleDesc.Quality = 0u;
     swap_desc.BufferUsage = DXGI_USAGE_RENDER_TARGET_OUTPUT;
-    swap_desc.BufferCount = desc->buffer_count;
-    swap_desc.Scaling = DXGI_SCALING_STRETCH;
+    swap_desc.BufferCount = desc.buffer_count;
+    swap_desc.Scaling = convert_dx(desc.scaling);
     swap_desc.SwapEffect = DXGI_SWAP_EFFECT_FLIP_DISCARD;
     swap_desc.AlphaMode = DXGI_ALPHA_MODE_UNSPECIFIED;
     swap_desc.Flags = DXGI_SWAP_CHAIN_FLAG_FRAME_LATENCY_WAITABLE_OBJECT;
@@ -39,7 +39,7 @@ inline wis::com_ptr<ID3D11Device> CreateD3D11Device() noexcept
 } // namespace wis::detail
 
 wis::DX12SwapChain
-wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const DX12Device& device, DX12QueueView main_queue, const wis::SwapchainDesc* desc, HWND hwnd) const noexcept
+wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const DX12Device& device, DX12QueueView main_queue, const wis::SwapchainDesc& desc, HWND hwnd) const noexcept
 {
     DX12SwapChain out_swapchain;
     auto& internal = out_swapchain.GetMutableInternal();
@@ -59,7 +59,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const 
         factory5->CheckFeatureSupport(DXGI_FEATURE_PRESENT_ALLOW_TEARING, &xtearing, sizeof(xtearing));
         return bool(xtearing);
     }();
-    if (tearing && desc->tearing) {
+    if (tearing && desc.tearing) {
         swap_desc.Flags |= DXGI_SWAP_CHAIN_FLAG_ALLOW_TEARING;
     }
 
@@ -67,7 +67,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const 
     wis::com_ptr<IDXGISwapChain1> swap;
 
     // until microsoft fixes this
-    if (desc->stereo && !wis::succeeded(hr = devicei.factory->CreateSwapChainForHwnd(detail::CreateD3D11Device().get(), hwnd, &swap_desc, nullptr, nullptr, swap.put()))) {
+    if (desc.stereo && !wis::succeeded(hr = devicei.factory->CreateSwapChainForHwnd(detail::CreateD3D11Device().get(), hwnd, &swap_desc, nullptr, nullptr, swap.put()))) {
         result = wis::make_result<FUNC, "Failed to create D3D11 device for stereo mode">(hr);
     }
 
@@ -91,8 +91,8 @@ wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const 
 
     internal.present_event = internal.chain->GetFrameLatencyWaitableObject();
     internal.stereo = swap_desc.Stereo;
-    internal.vsync = desc->vsync;
-    internal.tearing = tearing && desc->tearing;
+    internal.vsync = desc.vsync;
+    internal.tearing = tearing && desc.tearing;
 
     if (auto resw = internal.InitBackBuffers(); resw.status != wis::Status::Ok) {
         result = resw;
@@ -102,7 +102,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchain(wis::Result& result, const 
 }
 
 wis::DX12SwapChain
-wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, const DX12Device& device, DX12QueueView main_queue, const wis::SwapchainDesc* desc, IUnknown* window) const noexcept
+wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, const DX12Device& device, DX12QueueView main_queue, const wis::SwapchainDesc& desc, IUnknown* window) const noexcept
 {
     DX12SwapChain out_swapchain;
     auto& internal = out_swapchain.GetMutableInternal();
@@ -122,7 +122,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, con
         factory5->CheckFeatureSupport(DXGI_FEATURE_PRESENT_ALLOW_TEARING, &xtearing, sizeof(xtearing));
         return bool(xtearing);
     }();
-    if (tearing && desc->tearing) {
+    if (tearing && desc.tearing) {
         swap_desc.Flags |= DXGI_SWAP_CHAIN_FLAG_ALLOW_TEARING;
     }
 
@@ -130,7 +130,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, con
     wis::com_ptr<IDXGISwapChain1> swap;
 
     // until microsoft fixes this
-    if (desc->stereo && !wis::succeeded(hr = devicei.factory->CreateSwapChainForCoreWindow(detail::CreateD3D11Device().get(), window, &swap_desc, nullptr, swap.put()))) {
+    if (desc.stereo && !wis::succeeded(hr = devicei.factory->CreateSwapChainForCoreWindow(detail::CreateD3D11Device().get(), window, &swap_desc, nullptr, swap.put()))) {
         result = wis::make_result<FUNC, "Failed to create D3D11 device for stereo mode">(hr);
     }
 
@@ -153,8 +153,8 @@ wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, con
 
     internal.present_event = internal.chain->GetFrameLatencyWaitableObject();
     internal.stereo = swap_desc.Stereo;
-    internal.vsync = desc->vsync;
-    internal.tearing = tearing && desc->tearing;
+    internal.vsync = desc.vsync;
+    internal.tearing = tearing && desc.tearing;
 
     if (auto resw = internal.InitBackBuffers(); resw.status != wis::Status::Ok) {
         result = resw;
@@ -169,7 +169,7 @@ wis::platform::DX12WindowsExtension::CreateSwapchainUWP(wis::Result& result, con
 // #error error
 
 wis::VKSwapChain
-wis::platform::VKWindowsExtension::CreateSwapchain(wis::Result& result, const VKDevice& device, VKQueueView main_queue, const wis::SwapchainDesc* desc, HWND hwnd) const noexcept
+wis::platform::VKWindowsExtension::CreateSwapchain(wis::Result& result, const VKDevice& device, VKQueueView main_queue, const wis::SwapchainDesc& desc, HWND hwnd) const noexcept
 {
     VkWin32SurfaceCreateInfoKHR surface_desc{
         .sType = VK_STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR,
@@ -188,6 +188,8 @@ wis::platform::VKWindowsExtension::CreateSwapchain(wis::Result& result, const VK
         result = wis::make_result<FUNC, "Failed to create Win32 surface">(vr);
         return {};
     }
+
+
     wis::SharedSurface surface_handle{ surface, instance, instance_table.vkDestroySurfaceKHR };
     return device.VKCreateSwapChain(result, surface_handle, desc, std::get<0>(main_queue));
 }